refactor: rename library to templates
