+++ apps/spaces/src/components/QuestionLibraryListLayout/components/QuestionTemplateResults.tsx
@@ -1,17 +1,11 @@
-import { ComponentProps, Dispatch, FunctionComponent, SetStateAction, useEffect, useState } from "react";
-import { CardPagination, Table } from "@horizontal-org/shira-ui";
+import { Dispatch, FunctionComponent, SetStateAction } from "react";
+import { Table } from "@horizontal-org/shira-ui";
 import { ColumnDef, RowSelectionState } from "@tanstack/react-table";
 import { useTranslation } from "react-i18next";
 import { LibrarySearchEmptyState } from "../../LibrarySearchEmptyState";
 import { RowType } from "./Columns";
-import {
-  InactiveLibraryPaginationContainer,
-  LibraryPaginationContainer,
-} from "../../TemplatePaginationWrapper";
 
 type Props = {
-  shouldShowPagination: boolean;
-  paginationProps: ComponentProps<typeof CardPagination>;
   showEmptyState: boolean;
   loading: boolean;
   rows: RowType[];
@@ -21,8 +15,6 @@ type Props = {
 };
 
 export const QuestionTemplateResults: FunctionComponent<Props> = ({
-  shouldShowPagination,
-  paginationProps,
   showEmptyState,
   loading,
   rows,
@@ -31,30 +23,9 @@ export const QuestionTemplateResults: FunctionComponent<Props> = ({
   setRowSelection,
 }) => {
   const { t } = useTranslation();
-  const [stablePaginationProps, setStablePaginationProps] = useState(paginationProps);
-
-  useEffect(() => {
-    if (shouldShowPagination) {
-      setStablePaginationProps(paginationProps);
-    }
-  }, [paginationProps, shouldShowPagination]);
-
-  const shouldKeepPaginationVisible = rows.length > 0 && (loading || showEmptyState);
-  const paginationPropsToRender = shouldShowPagination ? paginationProps : stablePaginationProps;
-  const shouldShowTableLoadingState = loading && rows.length === 0;
 
   return (
     <>
-      {shouldShowPagination ? (
-        <LibraryPaginationContainer>
-          <CardPagination {...paginationPropsToRender} />
-        </LibraryPaginationContainer>
-      ) : shouldKeepPaginationVisible ? (
-        <InactiveLibraryPaginationContainer>
-          <CardPagination {...paginationPropsToRender} />
-        </InactiveLibraryPaginationContainer>
-      ) : null}
-
       {showEmptyState ? (
         <LibrarySearchEmptyState
           title={t("library.empty_search.title")}
@@ -63,13 +34,13 @@ export const QuestionTemplateResults: FunctionComponent<Props> = ({
       ) : (
         <Table
           size="full"
-          loading={shouldShowTableLoadingState}
+          loading={loading}
           data={rows}
           columns={columns}
           rowSelection={rowSelection}
           setRowSelection={setRowSelection}
           enableRowSelection={false}
-          enablePagination={false}
+          enablePagination={true}
           colGroups={
             <colgroup>
               <col style={{ width: "4%" }} />
@@ -84,16 +55,6 @@ export const QuestionTemplateResults: FunctionComponent<Props> = ({
           }
         />
       )}
-
-      {shouldShowPagination ? (
-        <LibraryPaginationContainer>
-          <CardPagination {...paginationPropsToRender} />
-        </LibraryPaginationContainer>
-      ) : shouldKeepPaginationVisible ? (
-        <InactiveLibraryPaginationContainer>
-          <CardPagination {...paginationPropsToRender} />
-        </InactiveLibraryPaginationContainer>
-      ) : null}
     </>
   );
 };
