+++ apps/spaces/src/components/QuestionLibraryListLayout/hooks/useQuestionTemplateList.ts
@@ -1,4 +1,4 @@
-import { type SetStateAction, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { getApps, type App } from "../../../fetch/app";
 import {
@@ -9,7 +9,6 @@ import {
   type QuestionTemplateFilterOption,
   type QuestionTemplateSortOption,
 } from "../../../fetch/question_templates";
-import { usePaginationProps } from "../../../hooks/usePaginationProps";
 import { useDebouncedValue } from "./useDebouncedValue";
 import { useQuestionTemplateFilterOptions } from "./useQuestionTemplateFilterOptions";
 import { useQuestionTemplateFilters } from "./useQuestionTemplateFilters";
@@ -24,21 +23,12 @@ export const useQuestionTemplateList = () => {
   const [searchValue, setSearchValueState] = useState("");
   const [sortOption, setSortOptionState] = useState<QuestionTemplateSortOption>(DEFAULT_QUESTION_TEMPLATE_SORT);
 
-  const startLoading = () => {
-    setLoading(true);
-  };
-
-  const setPageIndex = (value: SetStateAction<number>) => {
-    startLoading();
-    setPageIndexState(value);
-  };
-
-  const resetPagination = () => setPageIndexState(0);
+const resetPagination = () => setPageIndexState(0);
 
   const debouncedSearchValue = useDebouncedValue(searchValue.trim(), 300);
 
   const filters = useQuestionTemplateFilters(() => {
-    startLoading();
+    setLoading(true);
     resetPagination();
   });
   const { languageOptions, tagOptions } = useQuestionTemplateFilterOptions(filters.areFiltersOpen);
@@ -126,24 +116,16 @@ export const useQuestionTemplateList = () => {
   }, [pageCount, pageIndex]);
 
   const setSearchValue = (value: string) => {
-    startLoading();
+    setLoading(true);
     setSearchValueState(value);
     resetPagination();
   };
 
   const setSortOption = (nextSortOption: QuestionTemplateSortOption) => {
-    startLoading();
+    setLoading(false);
     setSortOptionState(nextSortOption);
     resetPagination();
   };
-
-  const paginationProps = usePaginationProps({
-    pageIndex,
-    pageCount,
-    pageSize: DEFAULT_PAGE_LIMIT,
-    setPageIndex,
-    total,
-  });
 
   return {
     ...filters,
@@ -154,7 +136,6 @@ export const useQuestionTemplateList = () => {
     languageOptions,
     loading,
     pageIndex,
-    paginationProps,
     questionTemplates,
     searchValue,
     setSearchValue,
